Fp16

+++ plot_lora.py
@@ -1,4 +1,5 @@
 import logging
+import torch
 
 def log_lora(lora_layers, log_weights=True, log_grad=True, log_level=logging.INFO):
     if not log_weights and not log_grad:
@@ -18,15 +19,15 @@ def log_lora(lora_layers, log_weights=True, log_grad=True, log_level=logging.INF
         q = lora['q_lora']
         v = lora['v_lora']
         if log_grad:
-            gradients_a[f'Q{i}.A'] = q.A.weight.grad.view(-1).tolist()
-            gradients_b[f'Q{i}.B'] = q.B.weight.grad.view(-1).tolist()
-            gradients_a[f'V{i}.A'] = v.A.weight.grad.view(-1).tolist()
-            gradients_b[f'V{i}.B'] = v.B.weight.grad.view(-1).tolist()
+            gradients_a[f'Q{i}.A'] = q.A.weight.grad.view(-1).to(torch.float32).tolist()
+            gradients_b[f'Q{i}.B'] = q.B.weight.grad.view(-1).to(torch.float32).tolist()
+            gradients_a[f'V{i}.A'] = v.A.weight.grad.view(-1).to(torch.float32).tolist()
+            gradients_b[f'V{i}.B'] = v.B.weight.grad.view(-1).to(torch.float32).tolist()
         if log_weights:
-            weights_a[f'Q{i}.A'] = q.A.weight.view(-1).tolist()
-            weights_b[f'Q{i}.B'] = q.B.weight.view(-1).tolist()
-            weights_a[f'V{i}.A'] = v.A.weight.view(-1).tolist()
-            weights_b[f'V{i}.B'] = v.B.weight.view(-1).tolist()
+            weights_a[f'Q{i}.A'] = q.A.weight.view(-1).to(torch.float32).tolist()
+            weights_b[f'Q{i}.B'] = q.B.weight.view(-1).to(torch.float32).tolist()
+            weights_a[f'V{i}.A'] = v.A.weight.view(-1).to(torch.float32).tolist()
+            weights_b[f'V{i}.B'] = v.B.weight.view(-1).to(torch.float32).tolist()
 
     if log_grad:
         logging.log(log_level, f'\n=== GRADIENTS A ===')
